+++ warp/_src/jax/custom_call.py
@@ -25,11 +25,11 @@ def jax_kernel(kernel, launch_dims=None, quiet=False):
 
     .. deprecated:: 1.10.0
         This version of ``jax_kernel()`` is deprecated for JAX >= 0.5.0 and is not supported
-        with JAX >= 0.8.0. Use :func:`warp.jax.ffi.jax_kernel` instead, which
+        with JAX >= 0.8.0. Use :func:`warp.jax_kernel` instead, which
         is the default implementation as of Warp 1.10.
 
     This implementation requires JAX version 0.4.25 - 0.7.x. For JAX 0.8.0 and later,
-    use the FFI-based implementation at :func:`warp.jax.ffi.jax_kernel`.
+    use the FFI-based implementation at :func:`warp.jax_kernel`.
 
     Args:
         kernel: The Warp kernel to be wrapped.
@@ -56,14 +56,14 @@ def jax_kernel(kernel, launch_dims=None, quiet=False):
             f"but installed JAX version is {jax.__version_info__}."
         )
         if jax.__version_info__ >= (0, 8, 0):
-            msg += " Please use warp.jax.ffi.jax_kernel instead."
+            msg += " Please use warp.jax_kernel instead."
         raise RuntimeError(msg)
 
     # deprecation warning
     if jax.__version_info__ >= (0, 5, 0) and not quiet:
         log_warning(
             "This version of jax_kernel() is deprecated and will not be supported with newer JAX versions. "
-            "Please use the newer FFI version instead (warp.jax.ffi.jax_kernel). "
+            "Please use the newer FFI version instead (warp.jax_kernel). "
             "As of Warp release 1.10, the FFI version is the default implementation of jax_kernel(). "
             "Pass quiet=True to disable this warning.",
             DeprecationWarning,