+++ warp/examples/interop/example_jax_callable.py
@@ -13,7 +13,7 @@ import jax
 import jax.numpy as jnp
 
 import warp as wp
-from warp.jax import jax_callable
+from warp import jax_callable
 
 
 @wp.kernel